@@ -1,76 +1,31 @@
 {
-  "name": "react-boilerplate",
-  "version": "0.1.0",
+  "name": "y",
   "private": true,
-  "dependencies": {
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
-    "@mui/material": "^5.14.18",
-    "@mui/x-data-grid": "^6.18.2",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "axios": "^1.6.2",
-    "bootstrap": "^4.6.2",
-    "countries": "^1.0.2",
-    "crypto-js": "^4.1.1",
-    "firebase": "^8.10.0",
-    "iso-639-1": "^3.1.0",
-    "languages": "^0.1.3",
-    "react": "^17.0.2",
-    "react-apexcharts": "^1.4.1",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.12.0",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "5.0.1",
-    "react-select": "^5.8.0",
-    "react-toastify": "^9.1.3",
-    "reacticons": "^0.0.1",
-    "reactstrap": "^8.10.1",
-    "redux": "^4.1.0",
-    "redux-thunk": "^2.3.0",
-    "web-vitals": "^1.0.1",
-    "yarn": "^1.22.21"
-  },
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "prepare": "husky install"
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "dependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^5.0.1",
+    "react-router-dom": "^6.22.3"
   },
   "devDependencies": {
-    "husky": "^8.0.2",
-    "lint-staged": "^13.0.4",
-    "prettier": "^2.8.0",
-    "tailwindcss": "^3.4.0",
-    "typescript": "^5.3.3"
-  },
-  "lint-staged": {
-    "*.{js,css,md}": "prettier --write"
-  },
-  "description": "All necessary files setup completed and ready to start.\r **Kindly follow the structure**",
-  "main": "postcss.config.js",
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
+    "@types/react": "^18.2.56",
+    "@types/react-dom": "^18.2.19",
+    "@vitejs/plugin-react": "^4.2.1",
+    "autoprefixer": "^10.4.18",
+    "eslint": "^8.56.0",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.5",
+    "postcss": "^8.4.35",
+    "tailwindcss": "^3.4.1",
+    "vite": "^5.1.4"
+  }
 }
