@@ -1,10 +1,16 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: ["./src/**/*.{js,jsx,ts,tsx}",
-  "./public/index.html"],
+export default {
+  content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
   theme: {
-    extend: {},
+    extend: {
+      screens: {
+        md1: "836px",
+        XL: "1280px",
+        md2: "920px",
+        "custom-md": "840px",
+        "custom-lg": "1200px",
+      },
+    },
   },
   plugins: [],
-}
-
+};
